Fix tuple types (#2634)

+++ wrappers/Python/CoolProp/CoolProp.pyx
@@ -143,7 +143,7 @@ cdef class ChebyshevExpansion:
     def solve_for_x(self, double y, double a, double b, unsigned int bits, size_t max_iter, double boundstytol):
         return self.m_exp.solve_for_x(y, a, b, bits, max_iter, boundstytol)
 
-    def solve_for_x_many(self, double[::1]  y, double a, double b, unsigned int bits, size_t max_iter, double boundstytol, double[::1] x, double [::1] counts):
+    def solve_for_x_many(self, double[::1]  y, double a, double b, unsigned int bits, size_t max_iter, double boundstytol, double[::1] x, size_t [::1] counts):
         cdef size_t N = y.shape[0]
         return self.m_exp.solve_for_x_manyC(&y[0], N, a, b, bits, max_iter, boundstytol, &x[0], &counts[0])
 
@@ -176,7 +176,7 @@ cdef class ChebyshevApproximation1D:
     def get_x_for_y(self, double y, unsigned int bits, size_t max_iter, double boundstytol):
         return self.thisptr.get_x_for_y(y, bits, max_iter, boundstytol)
 
-    def count_x_for_y_many(self, double[::1] y, unsigned int bits, size_t max_iter, double boundstytol, double[::1] counts):
+    def count_x_for_y_many(self, double[::1] y, unsigned int bits, size_t max_iter, double boundstytol, size_t[::1] counts):
         assert y.shape[0] == counts.shape[0]
         cdef size_t N = y.shape[0]
         return self.thisptr.count_x_for_y_manyC(&y[0], N, bits, max_iter, boundstytol, &counts[0])
